DL/app/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.46
botocore==1.35.46
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
fastapi==0.115.3
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
motor==3.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pymongo==4.9.2
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
sympy==1.13.1
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.20
ultralytics-thop==2.0.9
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
